requirements.txt
certifi==2023.5.7
chardet==4.0.0
charset-normalizer==3.2.0
cmake==3.27.0
filelock==3.12.2
fsspec==2023.6.0
geotorch==0.3.0
huggingface-hub==0.16.4
idna==2.10
Jinja2==2.11.3
joblib==1.3.1
lit==16.0.6
littleutils==0.2.2
MarkupSafe==2.1.3
matplotlib==3.4.2
mpmath==1.3.0
networkx==3.1
numpy==1.19.5
ogb==1.3.6
outdated==0.2.2
packaging==23.1
pandas==1.2.5
parameterized==0.9.0
Pillow==10.0.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
regex==2023.6.3
requests==2.25.1
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sympy==1.12
threadpoolctl==3.2.0
timm==0.6.13
tokenizers==0.13.3
torch==1.11.0
torchdiffeq==0.2.3
torchvision==0.12.0
torchaudio==0.11.0
tqdm==4.56.2
transformers==4.31.0
triton==2.0.0
typing-extensions==4.7.1
uncertainty-calibration==0.1.4
urllib3==1.26.16
git+https://github.com/RobustBench/robustbench.git@v1.0
wandb==0.15.0
wilds==2.0.0
